Moved methods from rooms to Room
