+++ package.json
@@ -24,6 +24,8 @@
   },
   "homepage": "https://github.com/thomson42/adventure-game#readme",
   "devDependencies": {
+    "enzyme": "^2.9.1",
+    "enzyme-to-json": "^1.5.1",
     "react-test-renderer": "^15.6.1"
   }
 }
