+++ src/modules/Room.js
@@ -1,32 +1,20 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import roomData from './rooms';
 
-Room.propTypes = {
-  room: PropTypes.object,
+const rooms = roomData;
+const items = ['cell phone', 'hall pass', 'math homework', 'banana peel', 'toilet paper statue', 'incriminating photograph', 'greasy pizza box'];
+const maxItm = 3;
 
-};
+rooms.forEach(room => {
+  room.doors = room.doors.map(door => {
+    return rooms.find(r => r.key === door);
+  });
+});
 
+function placeRandom(item) {
+  let randomIndex = Math.floor((Math.random() * rooms.length));
+  rooms[randomIndex].items.length === maxItm ? placeRandom(item) : rooms[randomIndex].items.push(item);
+}
 
+items.forEach(item => placeRandom(item));
 
-// export function Movies({ movies }) {
-//   return (
-//     <div className="movieList">
-//       {movies.map(movie =>
-//         <div key={movie.imdbID}>
-//           <Movie movie={movie} />
-//         </div>
-//       )}
-//     </div>
-//   );
-// }
-
-// export function Movie({ movie }) {
-//   return (
-//       <div className="movieItem movieInfo">
-//         {movie.Title} - {movie.Year}
-//         <div className="movieInfo">
-//         <img src={movie.Poster} style={{width: 100}}/>
-//       </div>
-//       </div>
-//   );
-// }
+export default rooms;
